+++ store/index.js
@@ -119,6 +119,12 @@ export const state = () => ({
           default: ''
         },
         {
+          name: 'required',
+          type: 'Boolean',
+          description: 'Required state',
+          default: 'false'
+        },
+        {
           name: 'autofocus',
           type: 'Boolean',
           description: 'Autofocus state',
@@ -143,12 +149,117 @@ export const state = () => ({
   v-model="value"
   placeholder="placeholder"
   @change="method"
-  autofocus
 >
   Label text here
 </common-input>`
+    },
+    {
+      id: 'loader',
+      title: 'Loader',
+      properties: [
+        {
+          name: 'show',
+          type: 'Boolean',
+          description: 'Show state',
+          default: 'false'
+        }
+      ],
+      events: [],
+      example: '<common-loader :show="true" />'
+    },
+    {
+      id: 'textarea',
+      title: 'Textarea',
+      properties: [
+        {
+          name: 'id',
+          type: 'String',
+          description: 'ID for label and input tags',
+          default: 'random seven-digit number'
+        },
+        {
+          name: 'name',
+          type: 'String',
+          description: 'Input name value',
+          default: ''
+        },
+        {
+          name: 'type',
+          type: 'String',
+          description: 'Input type for custom styles',
+          default: 'default'
+        },
+        {
+          name: 'size',
+          type: 'String',
+          description: 'Input size for custom styles',
+          default: 'default'
+        },
+        {
+          name: 'v-model',
+          type: 'String, Number',
+          description: 'Reactive input value',
+          default: ''
+        },
+        {
+          name: 'error',
+          type: 'String, Boolean',
+          description: 'Error message or state',
+          default: 'false'
+        },
+        {
+          name: 'placeholder',
+          type: 'String',
+          description: 'Input placeholder text',
+          default: ''
+        },
+        {
+          name: 'cols',
+          type: 'Number',
+          description: 'Number of columns',
+          default: ''
+        },
+        {
+          name: 'rows',
+          type: 'Number',
+          description: 'Number of rows',
+          default: ''
+        },
+        {
+          name: 'required',
+          type: 'Boolean',
+          description: 'Required state',
+          default: 'false'
+        },
+        {
+          name: 'autofocus',
+          type: 'Boolean',
+          description: 'Autofocus state',
+          default: 'false'
+        },
+        {
+          name: 'readonly',
+          type: 'Boolean',
+          description: 'Readonly state',
+          default: 'false'
+        },
+        {
+          name: 'disabled',
+          type: 'Boolean',
+          description: 'Disabled state',
+          default: 'false'
+        }
+      ],
+      events: ['input', 'change', 'focus', 'blur', 'keyup', 'paste'],
+      example: `<common-textarea
+  id="id"
+  v-model="value"
+  :rows="3"
+  placeholder="placeholder"
+/>`
     }
-  ]
+  ],
+  inDev: ['checkbox', 'file', 'modal', 'radio', 'search', 'tabs']
 })
 
 export const getters = {